remove ci_worker_storage_driver variable
This variable was driving the Docker engine-storage-driver parameter.
We don't have a clear use case for allowing this to be set and in most
cases we're going to want Docker to set the default.  As currently
implemented the value was being set from our own variable default rather
than allowing Docker to set its own.  It can be re-introduced if a clear
use case emerges and if it can be implemented in a way where Docker can
choose the default.