Revert a couple of pcboy's changes I don't fully understand yet
pcboy has made a few changes to the startup script that may be required
but I don't understand them fully. I'll back them out for now until I
understand if and why they're required. This will allow the remainder of
the PR to be merged.

During startup the runner creates a test docker-machine to make sure the
config is working. The step to remove the test instance was changed run
in the background in case it hung. I have two concerns with this:
* I've never seen it before. Was it a one off?
* It makes the startup logs harder to diagnose due to interleaving of
activity.

The gitlab runner registration step was modified to use a "|| true"
instead of "&& true" which will make the command return success even if
it fails. I don't understand why this is required.